refactor: reorganize components
- update AppRouter.jsx
- update HomePage.jsx
- update HomePage.css
- update FavoritesPage.jsx
- update FavoritesPage.css
- update Favorites.jsx
- update Favorites.css
- update Favorite.jsx
- update Favorite.css
- update ReviewsPage.jsx
- update BlogsPage.jsx
- update AboutPage.jsx
