+++ src/components/FavoritesPage/Favorite.css
