+++ src/components/FavoritesPage/Favorites.css
