+++ src/components/FavoritesPage/Favorites.jsx
