+++ src/components/FavoritesPage/FavoritesPage.css
